refactor: YMD 자리 이동
