+++ jpa/src/main/kotlin/com/comento/jpa/domain/company/Company.kt
@@ -28,18 +28,6 @@ data class Company(
         get() = _id ?: 0
 }
 
-@JvmInline
-value class YMD(private val date: String) : Comparable<YMD> {
 
-    init {
-        val re = Regex("^\\d{4}\\-(0?[1-9]|1[012])\\-(0?[1-9]|[12][0-9]|3[01])$")
-        if (!re.matches(date)) throw IllegalArgumentException("Incorrect Format")
-    }
-
-    override fun compareTo(other: YMD): Int {
-        return date.compareTo(other.date)
-    }
-
-}
